script/packagecloud.rb: update distro releases
Ubuntu Impish Indri reached its EOL in July 2022, so we remove it
from the list of Debian 11 distributions for which we publish
packages to Packagecloud, and we add Ubuntu Kinetic Kudu to the
Debian 11 list as it was released in October 2022.

Fedora 34 reached its EOL in July 2022 and Fedora 35 reached its
EOL in December 2022, so we remove them from the list of CentOS 8
distributions, and we add Fedora 37 to the Rocky 9 list as it was
released in November 2022.

We add Linux Mint Vera to the list of Debian 11 distributions
as it was released in December 2022.

We remove OpenSUSE 15.3 from the list of CentOS 7 distributions
as it reached its EOL in December 2022.

SUSE Linux Enterprise Server 12 SP4 reached its general EOL
in June 2020, so we remove it now, and SLES 15 SP3 reached its
general EOL in December 2022, so we also remove it now.

Note that we have been somewhat inconsistent in our handling of
SLES 12 and 15 Service Packs, as we have only removed SLES 12
Service Packs once they reached their LTSS (Long Term Service Pack
Support) EOL, while we have removed SLES 15 SPs when they reached
their general EOLs.  Thus we removed SLES 12 SP3 in commit
9ef46f8d606a05d5a51ce024c38445f1a42ecb11 of PR #5169 when that
Service Pack reached its LTSS EOL, but we removed SLES 15 SP2
in commit e2465f84b50452ff746f4f389afe5f8052a5a658 PR #5010 when
that Service Pack reached its general EOL but not its LTSS EOL,
which is still in the future (December 2024), as is the LTSS EOL
for SLES 15 SP1 (January 2024).

To be more consistent, we assume SLES users have upgraded to
the latest Service Pack within the default six-month period
after its release, and we can drop support for older SPs at
that point.  This is effectively how we have already treated
the earlier SPs for SLES 15, so applying that guideline to
SLES 12 SP4 as well should be reasonable.