standalone: report errors when handler setup fails
When the handler setup fails, we just exit right now, instead of sending
an error message.  That leads to less than helpful behaviour if the
destination repository isn't a valid Git repository for whatever reason.

To improve this, let's make sure that we return an error message at
setup if something fails so that the user in turn gets an error message.