standalone: include stderr if `git rev-parse` fails
If `git rev-parse --git-dir` fails, Git doesn't think that the
destination is a valid repository for whatever reason.  One of the most
common reasons for that is that the repository is owned by another user.

In such a case, we want to show the error message from Git, since we
really have no other indication that this is the case.  It could be just
that the destination isn't a repository at all.  Regardless, the error
message would be useful to the user to determine the problem, so let's
return it if standard error isn't empty.