locks: ensure local locks are synced on error
On an error, we call `os.Exit(2)`.  However, the defer for the locking
client that saves the local locks is not called in such a case, which
can leave the locks out of sync even if some of the operation fails.
If we're going to exit early, be sure we close the lock client to avoid
this problem.