Update golang.org/x/text
The version we're using is vulnerable to CVE-2022-32149.  We don't use
the vulnerable code, but upgrade the remaining golang.org/x packages we
depend on, since the text package is merely a dependency of them.  In
doing so, we'll update the text package, and avoid triggering any
warning from vulnerability scanners.

In addition, now that we're using released versions, it will be easier
to keep things up to date.