Avoid prefixing paths without directory separators
In Go 1.20, `filepath.Rel` always produces a path starting with `.` on
Windows if the path is relative to the current directory.  While this is
valid, it's not helpful, because we want to trim off the leading
directory portion leaving us with a bare relative path.  If we accepted
this behaviour, we'd break our checking for already-present patterns in
`git lfs track`, and more importantly, we wouldn't recognize raw
patterns wherever they are in the tree, since we'd falsely believe that
those patterns were specified with a leading dot.

To fix this, always trim the current directory prefix from our relative
directory and explicitly check for `.` and turn it into an empty string.