docs/README.md: remove stale video link
In commit 2f5b120dba77dfb2e2d03422864d7462717f1347 of PR #682 a link
to a YouTube video on the "GitHub Training & Guides" channel was added
to the docs/README.md file.  The video was named "Git Large File
Storage - How to Work with Big Files".

However, this YouTube channel has now been decommissioned by GitHub
and the video is no longer available, so until a replacement is
available, we remove the link to the old video.