tasklog/log.go: simplify consume() method
The *Logger.consume() method in the "tasklog" package was first
introduced in commit e90d54dabb45fd61c0c98803e009961800422010 of
PR #2329, and since that time has included logic to manage an
internal "pending" list of tasks; however, that logic is never
exercised because tasks are only appended to the "pending" array
when the "next" variable is non-nil, but that variable is
initialized as nil and can only be set to another value if the
"pending" array is non-empty.

As a result, we can remove both variables and the related logic
and simplify the consume() method as a result.