tasklog/percentage_task.go: align receiver name
In commit e90d54dabb45fd61c0c98803e009961800422010 of PR #2329 the
"git/githistory/log" package was added, including the PercentageTask
structure and associated methods.  Then in commit
122b76543a494d634ffd8c0a9225e0459691ebaa of PR #2610 the Entry()
method was added to provide support for verbose logging in the
"git lfs migrate impot" command.  However, the name of the
receiver argument for the Entry() method differs from that of
the other methods for the PercentageTask structure, so we update
it to match now, as well as fixing a typo which remains from the
original commit in PR #2329.

Note that the "git/githistory/log" package was later renamed to the
"tasklog" package in commits b9ab79ef899c4ee7cb656cdcbed666eed04c3022
and 45c580e60abfa36d0f9affba8b236f0e6516ce1d of PR #2747.