git/githistory/rewriter.go: always complete task
In a prior commit in this PR, we added a Complete() method for
the PercentageTask structure which sets the number of completed
elements of the task to the task's expected total in an atomic swap,
and then closes the task's updates channel if the number of completed
elements was less than the expected total.

As noted in that same commit, the PercentageTask structure is used
in the methods of the Rewriter structure from the "git/githistory"
package.  It is initialized with the number of commits to be rewritten
during a "git lfs migrate" command's traversal of a Git repository's
history.  As each commit is rewritten, the Count() method is called to
increment the percentage completed value.  When every commit has been
rewritten, the count reaches the expected total, and the Count() method
closes the channel, thus allowing the task receiving updates to finish
and exit its goroutine.

Under exceptional circumstances, though, the Rewrite() method of
the Rewriter structure may never finish iterating through all
the expected commits, and return in error or via a panic call.
When this happens, the anonymous goroutine started by the consume()
method of the Logger structure in "tasklog" package that is waiting
on the PercentageTask's updates channel can never exit, leading to
a hung process which never fully exits.

We can resolve this problem now by defining a deferred call to the
PercentageTask's new Complete() method after we create the task
structure in the Rewrite() method of the Rewriter structure.
As a result, even if the Rewrite() method suffers a panic() or
other error condition causing an early return, the PercentageTask's
updates channel is closed and the waiting goroutine will complete
and exit normally.

This usage corresponds to the resource management approach used
elsewhere, such as in the scanAll() and pointersToFetchForRefs()
functions of the "git lfs fetch" command, where SimpleTask objects
are created and a deferred call to their Complete() method is
immediately established.  This technique is also used in the
UpdateRefs() method of the refUpdater structure in the "git/githistory"
package, where a ListTask is created and a matching deferred call
to its Complete() method is then defined.