commands/command_{fetch,prune}.go: tidy task setup
The SimpleTask structure of the "tasklog" package, which was introduced
in commit 7a760b6b43d1131bbac89a552cf656b7fa6ce17a of PR #2756, is used
at several points in the implementation of the "git lfs fetch" and
"git lfs prune" commands, and in each case is constructed with a
call to the NewSimpleTask() function, and then passed to a Logger
structure's Enqueue() method.

However, in commit 0cad488c23350ff35f969980753bf9798071942e of PR #2767,
a Logger.Simple() method was added to both create a new SimpleTask
structure and pass it to a Logger structure's Enqueue() method.  This
Logger.Simple() method was added to be consistent with other existing
similar methods such as Logger.List() and Logger.Percentage().

These other methods are used consistently throughout our code
to create and enqueue tasks, so we now update the locations where
we call NewSimpleTask() to use the Logger.Simple() method instead.

This also provides an opportunity to tidy some of the surrounding
code and whitespace.