commands/command_prune.go: always complete task
In a prior commit in this PR, we added a Complete() method for
the PercentageTask structure which sets the number of completed
elements of the task to the task's expected total in an atomic swap,
and then closes the task's updates channel if the number of completed
elements was less than the expected total.

As noted in that same commit, this allows us to establish a deferred
call to the Complete() method after creating a PercentageTask structure,
akin to how we handle other "tasklog" package task structures such as
ListTask and WaiterTask.  Use of this idiom ensures that even if the
calling function returns due to an unexpected error or panic condition,
the anonymous goroutine started by the consume() method of the Logger
structure that is waiting on the task's updates channel will exit
cleanly, preventing a hung process.

We therefore now update the other use of the PercentageTask structure,
in the "git lfs prune" command's pruneDeleteFiles() function, to
follow the same approach as we have already implemented for the
Rewrite() method of the Rewriter structure in the "git/githistory"
package.