commands: use defer to always close tasks
In the implementation of several commands we create structures with
the Task interface from the "tasklog" package, and enqueue them to be
reported by an anonymous goroutine created by the Logger structure's
consume() method for each task.  These goroutines only exit when the
channel returned by the the Task's Updates() method is closed.

In order to ensure that we always close the channel associated with
each task, we refactor the creation and use of these tasks into small
utility functions.  We can then establish a deferred call to the
tasks' Complete() methods immediately after creating the tasks.
As these deferred calls will always be executed when returning from
the new utility functions, even under exceptional conditions, the
waiting goroutines will never fail to exit and thus cause the entire
program to hang.

This usage corresponds to the resource management approach used
elsewhere, such as in the scanAll() and pointersToFetchForRefs()
functions of the "git lfs fetch" command, where SimpleTask objects
are created and a deferred call to their Complete() method is
immediately established.  This technique is also used in the
UpdateRefs() method of the refUpdater structure in the "git/githistory"
package, where a ListTask is created and a matching deferred call
to its Complete() method is then defined, and in previous commits
in this PR we updated the locations in our code where we create
PercentageTask structures to follow the same pattern too.