t: refine migrate fixup test with no gitattributes
The "migrate info (--fixup, no .gitattributes)" test was added
to the t/t-migrate-info.sh test suite in commit
4800c5e9882f9cf1e271a855baeb32dbe1f767ec of PR #4501, when we
added --fixup option support to the "git lfs migrate info" command.

As we expect to introduce additional, similar tests in subsequent
commits in this PR, we rename the test to clarify that no fixup
action is expected in this case, and refactor the test's invocation
of the "git lfs migrate info --fixup" command so that it does not
run inside a pipeline.  This latter change ensures that if the
command fails, the test also fails due to the "set -e" shell option.

While this change to avoid the use of a pipeline is not critical for
this test, it will be important for the similar tests we add in
subsequent commits, because they will need to validate that the
"git lfs migrate info --fixup" command does not simply crash and
produce no output on stdout, which, if it occured in a pipeline,
would not be detected.  This is particularly true in these tests
because we expect the command to produce no output on stdout, which
it would not do both on success and if it panics, so we can not
distinguish these cases by examining the number of lines of output.

We then make a copy of this test in the t/t-migrate-fixup.sh test
suite as well, except using the "git lfs migrate import --fixup"
command, as we also expect to add similar tests to this suite in
subsequent commits.  Note that for this test it is critical that
we supply the --yes option to the command, to avoid it waiting
forever for input because it has detected the log file as a new
file in the Git working tree and therefore has prompted for
confirmation to proceed.

Finally, we remove one unneeded "git add" command from the
setup_single_local_branch_tracked_corrupt() test helper function,
which we refactor further in subsequent commits.