t: add future tests of migrate "lfs" macro support
In a previous commit in this PR we have refactored our handling of
parsed lines from .gitattributes files, and in doing so we have
updated the Applied() method of the Tree structure in the "git/gitattr"
package so that it has to perform a type check on each Line to
confirm that it implements the PatternLine interface, before applying
that Git attribute's match pattern to the method's file path argument.

This resolves the problem reported in #5332 where the
"git lfs migrate info --fixup" and "git lfs migrate import --fixup"
commands would panic when they encountered a .gitattributes file
with a non-pattern macro definition line.

However, these commands do not yet fully support macro attributes,
as they never create a MacroProcessor structure and thus never call
its ProcessLines() method on the "raw" lines parsed from each
.gitattributes file they encounter while traversing a repository's
Git commits and trees.

As a result, if a valid .gitattributes file exists in a repository
and defines a macro attribute for the "lfs" filter, and assigns
this macro attribute to a pattern which matches a file that is not
a proper Git LFS object, the two commands ignore this and do not
either report or convert the object into a Git LFS object as they
should.

While we expect to resolve this problem in a subsequent commit,
for the present time we simply add two tests which illustrate the
issue, but which we comment out as they would otherwise cause their
test suites to fail.