commands,lfs: remove GitScanner internal mutex
The GitScanner structure and its methods were introduced in PR #1670,
and in commit bdbca399c46f0447f08066ce53185009b3db90ec of that PR
the structure's internal sync.Mutex element was added to protect access to
the "remote" and "skippedRefs" elements, as well as the "closed" element.
The structure's methods always acquire the mutex before reading or writing
these elements, particularly the "remote" one.

This type of design is typical of other structures where we expect
concurrent access to the structure's data, such as when multiple
goroutines attempt to read and write the same data.

However, in practice the GitScanner structure's "remote" and "skippedRefs"
elements are only ever initialized once, and indeed only initialized in
one specific use case by the (*uploadContext).buildGitScanner() method
for the "git lfs push" and "git lfs pre-push" commands.  Meanwhile,
the "closed" element is only used in the Close() method, and this
method is only ever called once per structure by our existing code.

Moreover, other elements of the GitScanner structure which have been
added since its original introduction are set and retrieved without
locking the mutex, such as the FoundLockable and PotentialLockables
elements, which are checked and read directly in the scanRefsToChan()
function of the "lfs" package.

As it never the case that multiple goroutines need to read and write
to the "remote", "skippedRefs", and "closed" internal elements of the
GitScanner structure, we can remove the mutex and the associated code,
which simplifies the setup of the structure.