commands/command_fetch.go: drop unused GitScanner
In commit d5fd15225d0d31508c12042de174d2fdc5201b84 of PR #1745 the
"git lfs fetch" command, among others, was revised to set the Filter
element of the GitScanner structure, and no longer pass a common
GitScanner created in the main fetchCommand() function to various
other functions such as scanAll() and fetchAll().

As a result, the GitScanner structure created in the fetchCommand()
function was no longer used, but it was not fully removed at the time,
as was done in the "git lfs checkout" command's checkoutCommand()
function, for instance.  Therefore we can simply remove this unused
structure's initialization now in the fetchCommand() function.